OAK cleanup - Update: Codestyle: (i) indent from 2 spaces to 4, (ii) removed unused imports, (iii) cleaned up various codestyle warnings of various types. - Delete: oak/output.txt - Move: oak/oak_stuff.py -> backend/oak.py
