--- backend/oak/output.txt
@@ -1,13 +1,0 @@
-N3C:4221281 ! Atrial dilatation
-Oak speed test step 1   basic search for term completed in 0:00:00.001568
- * OUTGOING: N3C:3655971 Atrial cardiopathy
- * OUTGOING: N3C:4027255 Structural disorder of heart
-Oak speed test step 2   parents completed in 0:00:00.004531
- * INCOMING: N3C:4027255 Structural disorder of heart
- * INCOMING: N3C:4027255 Structural disorder of heart
- * INCOMING: N3C:4027255 Structural disorder of heart
- * INCOMING: N3C:4027255 Structural disorder of heart
- * INCOMING: N3C:4027255 Structural disorder of heart
- * INCOMING: N3C:4027255 Structural disorder of heart
-Oak speed test step 3   children completed in 0:00:01.818300
-Oak speed test completed in 0:00:01.824399
